(#10): 타입 정의를 정리하고 ApiResponse 타입을 src/pages/types.d.ts로 이동
